requirements.txt
async-timeout==3.0.1
attrs==19.3.0
Babel==2.8.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
distlib==0.3.1
et-xmlfile==1.0.1
filelock==3.0.12
idna==2.9
importlib-metadata==1.7.0
integer==1.0.3
jdcal==1.4.1
Jinja2==2.11.2
MarkupSafe==1.1.1
multidict==4.7.6
mysql-connector-python==8.0.21
openpyxl==3.0.4
pipenv==2020.6.2
protobuf==3.12.2
pycparser==2.20
pyTelegramBotAPI==3.7.2
pytz==2020.1
PyYAML==5.3.1
soupsieve==2.0.1
SQLAlchemy==1.3.17
urllib3==1.25.9
uuid==1.30
